Fix delay launching screensaver on Nvidia graphics cards
On one Nvidia machine, the screensaver launches to a black screen and
takes several seconds to appear. This is super jarring. You
completely miss the start of the animation! It feels like it’s doing a
video mode change, even though it shouldn’t be. Regardless, we now
handle antialising inside the Flux library itself.